test: Disable "output file failure" test for unsupported filesystem
The "Failure to write output file" test assumes that the filesystem
supports read-only directories. Improve this by probing this assumption
before running the test.

Fixes #1150.

+++ test/suites/base.bash
@@ -1194,22 +1194,27 @@ EOF
     expect_stat cache_miss 1
 
     # -------------------------------------------------------------------------
-    TEST "Failure to write output file"
 
     mkdir dir
-
-    $CCACHE_COMPILE -c test1.c -o dir/test1.o
-    expect_stat preprocessed_cache_hit 0
-    expect_stat cache_miss 1
-    expect_stat bad_output_file 0
-
-    rm dir/test1.o
     chmod a-w dir
+    if ! touch dir/test 2>/dev/null; then
+        TEST "Failure to write output file"
 
-    $CCACHE_COMPILE -c test1.c -o dir/test1.o 2>/dev/null
-    expect_stat preprocessed_cache_hit 0
-    expect_stat cache_miss 1
-    expect_stat bad_output_file 1
+        mkdir dir
+
+        $CCACHE_COMPILE -c test1.c -o dir/test1.o
+        expect_stat preprocessed_cache_hit 0
+        expect_stat cache_miss 1
+        expect_stat bad_output_file 0
+
+        rm dir/test1.o
+        chmod a-w dir
+
+        $CCACHE_COMPILE -c test1.c -o dir/test1.o 2>/dev/null
+        expect_stat preprocessed_cache_hit 0
+        expect_stat cache_miss 1
+        expect_stat bad_output_file 1
+    fi
 
     # -------------------------------------------------------------------------
     TEST "Caching stderr"
